fix(trending): Correct trending_score type
- Update trending score fields from entities
- Modify SQL schemas to change trending_score from integer to real

+++ src/main/kotlin/me/loghub/api/entity/article/ArticleStats.kt
@@ -12,5 +12,5 @@ class ArticleStats(
     var commentCount: Int = 0,
 
     @Column(name = "trending_score", nullable = false)
-    var trendingScore: Int = 0,
+    var trendingScore: Double = 0.0,
 )
